Rename RunnerBase.ts and Refactor Execution Handling
Renamed RunnerBase.ts to runner.ts for convention consistency. Enhanced error handling and logging by introducing tool-specific debug messages when actions fail. Refactored GitVersion tool execution methods to use safeExecute for improved error management.
